+++ packages/dsv-charts/src/layouts/cartesian2.ts
@@ -1,33 +1,9 @@
 import { BaseLayout } from './base';
-
-export type Cartesian2LayoutConfigType = {
-  padding?: {
-    top?: number;
-    right?: number;
-    bottom?: number;
-    left?: number;
-  };
-};
-
-export type Cartesian2Rect = {
-  width: number;
-  height: number;
-  center: [number, number];
-  left: number;
-  right: number;
-  top: number;
-  bottom: number;
-};
-
-export type Cartesian2InnerRect = {
-  innerWidth: number;
-  innerHeight: number;
-  innerCenter: [number, number];
-  innerLeft: number;
-  innerRight: number;
-  innerTop: number;
-  innerBottom: number;
-};
+import {
+  Cartesian2LayoutConfigType,
+  Cartesian2Rect,
+  Cartesian2InnerRect,
+} from './type';
 
 export class Cartesian2Layout extends BaseLayout {
   private _cartesian2Rect: Cartesian2Rect;
