+++ packages/dsv-charts/src/layouts/index.ts
@@ -1,1 +1,3 @@
 export * from './cartesian2';
+export * from './type';
+export * from './base';
