+++ packages/dsv-charts/src/types/util.ts
@@ -1,0 +1,4 @@
+export type ArrayLike<T> = {
+  [key: number]: T;
+  length: number;
+};
