--- packages/dsv-charts/src/typings/config/charts/array.ts
@@ -1,9 +1,0 @@
-export type ArrayKeyType = 'array';
-
-export type ArrayItemType = {
-  key: string;
-  value: number;
-  name: string;
-};
-
-export type ArrayDataType = Array<ArrayItemType>;
