--- packages/dsv-charts/src/typings/config/charts/queue.ts
@@ -1,9 +1,0 @@
-export type QueueKeyType = 'queue';
-
-export type QueueItemType = {
-  key: string;
-  value: number;
-  name: string;
-};
-
-export type QueueDataType = Array<QueueItemType>;
