--- packages/dsv-charts/src/typings/config/charts/stack.ts
@@ -1,9 +1,0 @@
-export type StackKeyType = 'stack';
-
-export type StackItemType = {
-  key: string;
-  value: number;
-  name: string;
-};
-
-export type StackDataType = Array<StackItemType>;
