--- packages/dsv-charts/src/typings/config/layouts/cartesian2.ts
@@ -1,8 +1,0 @@
-export type Cartesian2LayoutType = {
-  padding?: {
-    top?: number;
-    right?: number;
-    bottom?: number;
-    left?: number;
-  };
-};
