--- packages/dsv-charts/src/typings/config/life-circle/index.ts
@@ -1,8 +1,0 @@
-export interface ILifeCircle {
-  beforeInit?: (...args) => void;
-  afterInit?: (...args) => void;
-  brforeSetData?: (...args) => void;
-  afterSetData?: (...args) => void;
-}
-
-export type ILifeCircleKeys = keyof ILifeCircle;
