--- packages/dsv-charts/src/typings/index.ts
@@ -1,3 +1,0 @@
-export * from './config';
-export * from './theme';
-
